fix: address all clang-tidy warnings from CI report
Real bugs fixed:
- logger_json.c: check ftell() return for -1 before using as malloc size;
  clamp bytes_read before buffer indexing; early-exit on empty file
- db_auth.c: remove dead store (address = trimmed, always overwritten)
- db_core.c: guard qsort() against NULL backup_paths; add explicit
  (void*) casts for multi-level pointer conversions on realloc/qsort/free
- db_backup.c: add ferror() check after fread loop to catch silent errors
- sqlite_migrate.c: clamp fread result before indexing content buffer
- storage_manager.c: check dirfd() != -1 before fstatat()
- hls_directory.c: check dirfd() != -1 before fstatat() in .ts and .m4s loops
- go2rtc_process.c: guard recursive_remove_at() against negative parent fd
- hls_unified_thread.c: release unified_contexts_mutex before sleep(3)
  and reacquire after to avoid blocking other threads in critical section
- api_handlers_system.c: fix fp leak when cpu_max_parsed && period==0;
  restructure snprintf to avoid branch-clone warning; guard /proc/net/dev
  header reads before the parse loop
- api_handlers_system_logs.c: clamp bytes_read before buffer indexing

Code quality:
- api_handlers_auth_backend_agnostic.c: explicit (double) casts for
  int64_t fields passed to cJSON_AddNumberToObject (narrowing warnings)
- api_handlers_recording_tags.c: explicit casts for const char** malloc/free
- api_handlers_timeline.c: eliminate dead store of sqlite3_step result
- onvif_device_management.c: NOLINT on intentional non-null-terminated memcpy
- stream_manager.c: explicit cap count <= streams_capacity
- config.c: explicit cap loaded <= max_streams; NOLINT tainted isspace
- motion_detection.c: NOLINT on bounds-checked array accesses in box blur
- api_handlers_recordings_batch_download.c: NOLINTNEXTLINE on fread loops
- api_handlers_system_logs_tail.c: NOLINTNEXTLINE on fgets loops
- go2rtc_process.c: NOLINTNEXTLINE on fgets loop; NOLINT on strcspn indexing
- api_handlers_system.c: NOLINT on strcspn MAC address indexing

+++ src/web/api_handlers_auth_backend_agnostic.c
@@ -549,7 +549,7 @@ void handle_auth_verify(const http_request_t *req, http_response_t *res) {
         // Send success response with user info
         cJSON *response = cJSON_CreateObject();
         cJSON_AddBoolToObject(response, "authenticated", true);
-        cJSON_AddNumberToObject(response, "id", user.id);
+        cJSON_AddNumberToObject(response, "id", (double)user.id);
         cJSON_AddStringToObject(response, "username", user.username);
         cJSON_AddStringToObject(response, "email", user.email);
         cJSON_AddStringToObject(response, "role", db_auth_get_role_name(user.role));
@@ -632,11 +632,11 @@ void handle_auth_sessions_list(const http_request_t *req, http_response_t *res) 
     cJSON *items = cJSON_AddArrayToObject(response, "sessions");
     for (int i = 0; i < count; i++) {
         cJSON *item = cJSON_CreateObject();
-        cJSON_AddNumberToObject(item, "id", sessions[i].id);
-        cJSON_AddNumberToObject(item, "created_at", sessions[i].created_at);
-        cJSON_AddNumberToObject(item, "last_activity_at", sessions[i].last_activity_at);
-        cJSON_AddNumberToObject(item, "idle_expires_at", sessions[i].idle_expires_at);
-        cJSON_AddNumberToObject(item, "expires_at", sessions[i].expires_at);
+        cJSON_AddNumberToObject(item, "id", (double)sessions[i].id);
+        cJSON_AddNumberToObject(item, "created_at", (double)sessions[i].created_at);
+        cJSON_AddNumberToObject(item, "last_activity_at", (double)sessions[i].last_activity_at);
+        cJSON_AddNumberToObject(item, "idle_expires_at", (double)sessions[i].idle_expires_at);
+        cJSON_AddNumberToObject(item, "expires_at", (double)sessions[i].expires_at);
         cJSON_AddStringToObject(item, "ip_address", sessions[i].ip_address);
         cJSON_AddStringToObject(item, "user_agent", sessions[i].user_agent);
         cJSON_AddBoolToObject(item, "current", current_token[0] != '\0' && strcmp(current_token, sessions[i].token) == 0);
@@ -711,10 +711,10 @@ void handle_auth_trusted_devices_list(const http_request_t *req, http_response_t
     cJSON *items = cJSON_AddArrayToObject(response, "trusted_devices");
     for (int i = 0; i < count; i++) {
         cJSON *item = cJSON_CreateObject();
-        cJSON_AddNumberToObject(item, "id", devices[i].id);
-        cJSON_AddNumberToObject(item, "created_at", devices[i].created_at);
-        cJSON_AddNumberToObject(item, "last_used_at", devices[i].last_used_at);
-        cJSON_AddNumberToObject(item, "expires_at", devices[i].expires_at);
+        cJSON_AddNumberToObject(item, "id", (double)devices[i].id);
+        cJSON_AddNumberToObject(item, "created_at", (double)devices[i].created_at);
+        cJSON_AddNumberToObject(item, "last_used_at", (double)devices[i].last_used_at);
+        cJSON_AddNumberToObject(item, "expires_at", (double)devices[i].expires_at);
         cJSON_AddStringToObject(item, "ip_address", devices[i].ip_address);
         cJSON_AddStringToObject(item, "user_agent", devices[i].user_agent);
         cJSON_AddBoolToObject(item, "current", current_device_id > 0 && current_device_id == devices[i].id);
